Moved and renamed `PInvoke` -> `NativeMethods`.
